Move UdevRulesOracleASM from etc_udev_rules to udev_rules
